fix(jforms) improve mimetype validation of uploaded files
Security issue: uploaded file could be a php file disguised as an image.
The magic bytes AND the filename extension should be checked to not store
unexpected php files that could be executed.

+++ lib/jelix/forms/controls/jFormsControlImageUpload.class.php
@@ -88,18 +88,14 @@ class jFormsControlImageUpload extends jFormsControlUpload2
             return null;
         }
 
-        if (count($this->mimetype)) {
-            $this->fileInfo['type'] = \Jelix\FileUtilities\File::getMimeType($filePath);
-            if ($this->fileInfo['type'] == 'application/octet-stream') {
-                // let's try with the name
-                $this->fileInfo['type'] = \jFile::getMimeTypeFromFilename($this->fileInfo['name']);
+        if ($this->error === null && count($this->mimetype)) {
+            $mimeType = \jFile::verifyFileMimeType($filePath, $this->mimetype, $this->fileInfo['name']);
+            if ($mimeType) {
+                $this->fileInfo['type'] = $mimeType;
             }
-
-            if (!in_array($this->fileInfo['type'], $this->mimetype)) {
-                $this->error = \jForms::ERRDATA_INVALID_FILE_TYPE;
+            else {
+                $this->error = jForms::ERRDATA_INVALID_FILE_TYPE;
                 unlink($filePath);
-
-                return null;
             }
         }
